Rename Shared to Components; Complete renamte DataAccessLayer to Database
