+++ SpotifyAnalysis/Components/GridItems/MudItemArtistPane.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 @using MudBlazor
 @using Data.DTO
