+++ SpotifyAnalysis/Components/GridItems/MudItemBase.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 <MudItem xl="3" lg="6" md="12">
 	<MudPaper Class="full-height pa-4">
