+++ SpotifyAnalysis/Components/GridItems/MudItemPieChart.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 @using SpotifyAnalysis.Data
 @using ApexCharts
