+++ SpotifyAnalysis/Components/GridItems/WidgetArtists.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 @inherits WidgetBase
