+++ SpotifyAnalysis/Components/GridItems/WidgetGenres.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 @inherits WidgetBase
 
@@ -7,7 +7,7 @@
 
 @if (elements is not null) {
     @CreateChart()
-    <MudItemTracksPane Tracks=@genreTracks TitleSuffix=@genreName ShowGenres=true />
+    <MudItemTracksPane Tracks=@genreTracks TitleSuffix=@genreName />
 }
 
 @code {
@@ -15,9 +15,8 @@
     public IEnumerable<TrackDTO> Tracks { get; set; }
     [Parameter, EditorRequired]
     public IDictionary<string, ArtistDTO> ArtistsDict { get; set; }
-    public override string Title { get; } = "Genres General";
+    public override string Title { get; } = "Genres";
 
-    static readonly string[] excludes = { "of", "and", "new", "music" };
     protected override Elements BuildElements() {
         if (Tracks is null) return new Elements();
 
@@ -25,14 +24,11 @@
         foreach (var track in Tracks)
             foreach (var artist in track.Artists)
                 foreach (var genre in ArtistsDict[artist.ID].Genres)
-                    foreach (string word in genre.ToLower().Split(' ', StringSplitOptions.RemoveEmptyEntries))
-                        genres.Increase(new Element { Label = word, Quantity = 1 });
-        foreach (var exclusion in excludes)
-            genres.Extract(exclusion);
+                    genres.Increase(new Element { Label = genre, Quantity = 1 });
         return genres;
     }
 
-    public WidgetGenresGeneral() {
+    public WidgetGenres() {
         onClickCallback = FindGenreTracks;
     }
 
@@ -40,7 +36,7 @@
     IEnumerable<TrackDTO> genreTracks;
     void FindGenreTracks(Element element) {
         genreName = element.Label;
-        var genreArtists = ArtistsDict.Select(a => a.Value).Where(a => a.Genres.Any(g => g.Contains(genreName)));
+        var genreArtists = ArtistsDict.Select(a => a.Value).Where(a => a.Genres.Contains(genreName));
         genreTracks = Tracks.Where(t => t.Artists.Intersect(genreArtists).Any());
         StateHasChanged();
     }
