+++ SpotifyAnalysis/Components/GridItems/WidgetGenresGeneral.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 @inherits WidgetBase
 
@@ -7,7 +7,7 @@
 
 @if (elements is not null) {
     @CreateChart()
-    <MudItemTracksPane Tracks=@genreTracks TitleSuffix=@genreName />
+    <MudItemTracksPane Tracks=@genreTracks TitleSuffix=@genreName ShowGenres=true />
 }
 
 @code {
@@ -15,8 +15,9 @@
     public IEnumerable<TrackDTO> Tracks { get; set; }
     [Parameter, EditorRequired]
     public IDictionary<string, ArtistDTO> ArtistsDict { get; set; }
-    public override string Title { get; } = "Genres";
+    public override string Title { get; } = "Genres General";
 
+    static readonly string[] excludes = { "of", "and", "new", "music" };
     protected override Elements BuildElements() {
         if (Tracks is null) return new Elements();
 
@@ -24,11 +25,14 @@
         foreach (var track in Tracks)
             foreach (var artist in track.Artists)
                 foreach (var genre in ArtistsDict[artist.ID].Genres)
-                    genres.Increase(new Element { Label = genre, Quantity = 1 });
+                    foreach (string word in genre.ToLower().Split(' ', StringSplitOptions.RemoveEmptyEntries))
+                        genres.Increase(new Element { Label = word, Quantity = 1 });
+        foreach (var exclusion in excludes)
+            genres.Extract(exclusion);
         return genres;
     }
 
-    public WidgetGenres() {
+    public WidgetGenresGeneral() {
         onClickCallback = FindGenreTracks;
     }
 
@@ -36,7 +40,7 @@
     IEnumerable<TrackDTO> genreTracks;
     void FindGenreTracks(Element element) {
         genreName = element.Label;
-        var genreArtists = ArtistsDict.Select(a => a.Value).Where(a => a.Genres.Contains(genreName));
+        var genreArtists = ArtistsDict.Select(a => a.Value).Where(a => a.Genres.Any(g => g.Contains(genreName)));
         genreTracks = Tracks.Where(t => t.Artists.Intersect(genreArtists).Any());
         StateHasChanged();
     }
