+++ SpotifyAnalysis/Components/Help/HelpButton.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 @inject IDialogService Dialog
