+++ SpotifyAnalysis/Components/Help/HelpDialog.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 <MudDialog >
     <DialogContent >
