+++ SpotifyAnalysis/Components/Help/HelpSearchPlaylist.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 
 @inject SpotifyMudTheme SpotifyTheme
