+++ SpotifyAnalysis/Components/Main/MainLayout.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 @inherits LayoutComponentBase
 
 @inject SpotifyMudTheme SpotifyTheme
