+++ SpotifyAnalysis/Components/Main/NavMenu.razor
@@ -1,4 +1,4 @@
-﻿@namespace SpotifyAnalysis.Shared
+﻿@namespace SpotifyAnalysis.Components
 @implements IDisposable
 
 @using SpotifyAnalysis.Data.DTO
