+++ SpotifyAnalysis/Pages/BrowseTracks.razor
@@ -2,7 +2,7 @@
 
 @using SpotifyAnalysis.Data
 @using SpotifyAnalysis.Data.DTO
-@using SpotifyAnalysis.Data.DataAccessLayer
+@using SpotifyAnalysis.Data.Database
 
 @inject SpotifyMudTheme SpotifyTheme
 @inject ScopedData data;
