+++ SpotifyAnalysis/Pages/SelectUser.razor
@@ -2,7 +2,7 @@
 
 @using SpotifyAnalysis.Data.SpotifyAPI
 @using SpotifyAnalysis.Data.DTO
-@using SpotifyAnalysis.Data.DataAccessLayer
+@using SpotifyAnalysis.Data.Database
 @using Microsoft.EntityFrameworkCore
 @using System.Text.RegularExpressions
