+++ SpotifyAnalysis/Startup.cs
@@ -5,11 +5,10 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using MudBlazor.Services;
-using SpotifyAnalysis.Data;
-using SpotifyAnalysis.Data.DataAccessLayer;
+using SpotifyAnalysis.Data.Database;
 using SpotifyAnalysis.Data.DTO;
 using SpotifyAnalysis.Data.SpotifyAPI;
-using SpotifyAnalysis.Shared;
+using SpotifyAnalysis.Components;
 using System;
 using System.Collections.Generic;
 using System.Linq;
