+++ SpotifyAnalysis/_Imports.razor
@@ -7,5 +7,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using SpotifyAnalysis
-@using SpotifyAnalysis.Shared
+@using SpotifyAnalysis.Components
 @using MudBlazor
